Refactor: extract CSS/style resolution into Style/StyleResolutionTrait (slice 5)
Pure, behavior-preserving move of the CSS / style-resolution concern out of
HtmlTransformer into a dedicated Style/StyleResolutionTrait. Methods extracted
verbatim (no logic or signature changes): presentationAttributes,
mergedPresentationStyle, isHighValueStyledElement, staticStyleRules,
safeVisualDeclarations, cssDeclarations, cssDeclarationString,
isSupportedCssSelector, matchesCssSelector, normalizeCssSelector,
matchesCssSelectorPart, presentationClassName, isBehaviorHookClassName,
layoutAttribute, hasGridLikeClass.

HtmlTransformer `use`s the new trait; $staticStyleRules property and the
class-promotion / provenance helpers stay on HtmlTransformer.

Refs #242 (slice 5: CSS/style resolution).
Parity 128->128, canonical + full composer test green, zero block-output diff.
