+++ php-transformer/src/HtmlToBlocks/Style/StyleResolutionTrait.php
@@ -1,0 +1,335 @@
+<?php
+
+namespace Automattic\BlocksEngine\PhpTransformer\HtmlToBlocks\Style;
+
+use DOMElement;
+
+/**
+ * CSS / style-resolution concern extracted from HtmlTransformer.
+ *
+ * Resolves an element's declared styling from the source `<style>`/linked CSS
+ * and computes presentation attributes: static CSS-rule collection, supported
+ * selector matching (`matchesCssSelector`), merged inline + matched-rule style
+ * (`mergedPresentationStyle`), CSS declaration parsing/serialization, layout
+ * attribute inference, and presentation class-name normalization. This is the
+ * CSS-rule resolution the font/typography path and `ButtonStyleResolver` rely
+ * on, given a single home so style work no longer collides in the god-object.
+ *
+ * Pure move: methods extracted verbatim from HtmlTransformer with no logic or
+ * signature changes. Methods reference `$this->attr()` / `$this->safeAnchor()`
+ * (DomHelpersTrait), `$this->promotedClassName()` / `$this->cardLikeChildCount()`,
+ * and the `$staticStyleRules` property, all composed onto HtmlTransformer.
+ */
+trait StyleResolutionTrait
+{
+    /**
+     * @return array<string, mixed>
+     */
+    private function presentationAttributes(DOMElement $element): array
+    {
+        $style = $this->mergedPresentationStyle($element);
+
+        return array_filter(array(
+            'anchor'    => $this->safeAnchor($this->attr($element, 'id')),
+            'className' => $this->promotedClassName($this->attr($element, 'class')),
+            'style'     => $style,
+            'layout'    => $this->layoutAttribute($element, $style),
+        ), static fn ($value): bool => is_array($value) ? array() !== $value : '' !== trim((string) $value));
+    }
+
+    private function mergedPresentationStyle(DOMElement $element): string
+    {
+        $inlineStyle = $this->attr($element, 'style');
+        if ( array() === $this->staticStyleRules || ! $this->isHighValueStyledElement($element) ) {
+            return $inlineStyle;
+        }
+
+        $declarations = array();
+        foreach ( $this->staticStyleRules as $rule ) {
+            if ( $this->matchesCssSelector($element, $rule['selector']) ) {
+                $declarations = array_merge($declarations, $rule['declarations']);
+            }
+        }
+
+        if ( array() === $declarations ) {
+            return $inlineStyle;
+        }
+
+        $declarations = array_merge($declarations, $this->cssDeclarations($inlineStyle));
+        return $this->cssDeclarationString($declarations);
+    }
+
+    private function isHighValueStyledElement(DOMElement $element): bool
+    {
+        $tagName = strtolower($element->tagName);
+        if ( in_array($tagName, array( 'button', 'nav', 'article' ), true) ) {
+            return true;
+        }
+
+        $tokens = strtolower(trim(implode(' ', array(
+            $this->attr($element, 'class'),
+            $this->attr($element, 'id'),
+            $this->attr($element, 'role'),
+        ))));
+
+        if ( preg_match('/(?:^|[^a-z0-9])(?:btn|button|cta|action|nav|menu|cards?|tile|panel|pricing|price|product|grid|columns|layout|stack|cluster|row|wrap|hero|masthead|banner|media|image|photo|gallery)(?:[^a-z0-9]|$)/', $tokens) ) {
+            return true;
+        }
+
+        if ( 'a' === $tagName ) {
+            for ( $node = $element->parentNode; $node instanceof DOMElement; $node = $node->parentNode ) {
+                $ancestorTokens = strtolower($this->attr($node, 'class') . ' ' . $this->attr($node, 'id'));
+                if ( preg_match('/(?:^|[^a-z0-9])(?:actions?|btns?|buttons?|cta|nav|menu|card|tile|panel|pricing|product)(?:[^a-z0-9]|$)/', $ancestorTokens) ) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * @return array<int, array{selector: string, declarations: array<string, string>}>
+     */
+    private function staticStyleRules(string $html, string $linkedCss): array
+    {
+        $css = trim($linkedCss);
+        if ( preg_match_all('@<style\b[^>]*>(.*?)</style>@is', $html, $matches) ) {
+            $css .= ( '' === $css ? '' : "\n" ) . implode("\n", array_map('trim', $matches[1]));
+        }
+
+        if ( '' === trim($css) ) {
+            return array();
+        }
+
+        $css = preg_replace('@/\*.*?\*/@s', '', $css) ?? $css;
+        $rules = array();
+        if ( ! preg_match_all('/([^{}]+)\{([^{}]+)\}/', $css, $matches, PREG_SET_ORDER) ) {
+            return array();
+        }
+
+        foreach ( $matches as $match ) {
+            $declarations = $this->safeVisualDeclarations($this->cssDeclarations((string) $match[2]));
+            if ( array() === $declarations ) {
+                continue;
+            }
+            foreach ( explode(',', (string) $match[1]) as $selector ) {
+                $selector = trim($selector);
+                if ( '' !== $selector && $this->isSupportedCssSelector($selector) ) {
+                    $rules[] = array(
+                        'selector' => $selector,
+                        'declarations' => $declarations,
+                    );
+                }
+            }
+        }
+
+        return array_slice($rules, 0, 200);
+    }
+
+    /**
+     * @param array<string, string> $declarations
+     * @return array<string, string>
+     */
+    private function safeVisualDeclarations(array $declarations): array
+    {
+        $safe = array_flip(array(
+            'background',
+            'background-color',
+            'background-image',
+            'border',
+            'border-color',
+            'border-radius',
+            'border-style',
+            'border-width',
+            'box-shadow',
+            'color',
+            'align-items',
+            'column-gap',
+            'display',
+            'flex-direction',
+            'flex-wrap',
+            'font-size',
+            'font-weight',
+            'gap',
+            'grid-template-columns',
+            'grid-template-rows',
+            'justify-content',
+            'line-height',
+            'margin',
+            'margin-bottom',
+            'margin-left',
+            'margin-right',
+            'margin-top',
+            'padding',
+            'padding-bottom',
+            'padding-left',
+            'padding-right',
+            'padding-top',
+            'row-gap',
+            'text-align',
+            'text-decoration',
+            'text-transform',
+        ));
+
+        return array_intersect_key($declarations, $safe);
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    private function cssDeclarations(string $style): array
+    {
+        $declarations = array();
+        foreach ( explode(';', $style) as $declaration ) {
+            if ( ! str_contains($declaration, ':') ) {
+                continue;
+            }
+            [$name, $value] = array_map('trim', explode(':', $declaration, 2));
+            $name = strtolower($name);
+            $value = preg_replace('/\s+/', ' ', $value) ?? $value;
+            $allowsImageUrl = in_array($name, array( 'background', 'background-image' ), true) && ! preg_match('/(?:expression\s*\(|javascript\s*:)/i', $value);
+            if ( '' !== $name && '' !== $value && ( $allowsImageUrl || ! preg_match('/(?:expression\s*\(|javascript\s*:|url\s*\()/i', $value) ) ) {
+                $declarations[$name] = $value;
+            }
+        }
+
+        return $declarations;
+    }
+
+    /**
+     * @param array<string, string> $declarations
+     */
+    private function cssDeclarationString(array $declarations): string
+    {
+        $parts = array();
+        foreach ( $declarations as $name => $value ) {
+            $parts[] = $name . ':' . $value;
+        }
+
+        return implode(';', $parts);
+    }
+
+    private function isSupportedCssSelector(string $selector): bool
+    {
+        $selector = $this->normalizeCssSelector($selector);
+        if ( '' === $selector || preg_match('/[>+~\[\]=]/', $selector) ) {
+            return false;
+        }
+
+        foreach ( preg_split('/\s+/', $selector) ?: array() as $part ) {
+            if ( ! preg_match('/^(?:[a-z][a-z0-9_-]*)?(?:\.[A-Za-z0-9_-]+)+$|^\.[A-Za-z0-9_-]+$/i', $part) ) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    private function matchesCssSelector(DOMElement $element, string $selector): bool
+    {
+        $parts = preg_split('/\s+/', $this->normalizeCssSelector($selector)) ?: array();
+        if ( array() === $parts ) {
+            return false;
+        }
+
+        if ( ! $this->matchesCssSelectorPart($element, $parts[count($parts) - 1]) ) {
+            return false;
+        }
+
+        $current = $element->parentNode instanceof DOMElement ? $element->parentNode : null;
+        for ( $index = count($parts) - 2; $index >= 0; --$index ) {
+            $matched = false;
+            for ( $node = $current; $node instanceof DOMElement; $node = $node->parentNode instanceof DOMElement ? $node->parentNode : null ) {
+                if ( $this->matchesCssSelectorPart($node, $parts[$index]) ) {
+                    $matched = true;
+                    $current = $node->parentNode instanceof DOMElement ? $node->parentNode : null;
+                    break;
+                }
+            }
+            if ( ! $matched ) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    private function normalizeCssSelector(string $selector): string
+    {
+        return trim(preg_replace('/:(?:hover|focus|active|visited|before|after|focus-visible)\b.*/', '', $selector) ?? $selector);
+    }
+
+    private function matchesCssSelectorPart(DOMElement $element, string $selector): bool
+    {
+        if ( ! preg_match('/^(?:(?<tag>[a-z][a-z0-9_-]*))?(?<classes>(?:\.[A-Za-z0-9_-]+)+)$/i', $selector, $match) ) {
+            return false;
+        }
+
+        if ( ! empty($match['tag']) && strtolower($match['tag']) !== strtolower($element->tagName) ) {
+            return false;
+        }
+
+        $classes = preg_split('/\./', ltrim((string) $match['classes'], '.')) ?: array();
+        $elementClasses = preg_split('/\s+/', trim($this->attr($element, 'class'))) ?: array();
+        foreach ( $classes as $class ) {
+            if ( ! in_array($class, $elementClasses, true) ) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    private function presentationClassName(string $className): string
+    {
+        $classes = preg_split('/\s+/', trim($className)) ?: array();
+        $classes = array_filter($classes, static fn (string $class): bool => '' !== $class && ! self::isBehaviorHookClassName($class));
+
+        return implode(' ', array_values(array_unique($classes)));
+    }
+
+    private static function isBehaviorHookClassName(string $className): bool
+    {
+        return 1 === preg_match('/^js(?:$|[-_:]|[A-Z])/', $className);
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    private function layoutAttribute(DOMElement $element, string $mergedStyle = ''): array
+    {
+        $declared = trim($this->attr($element, 'data-layout'));
+        if ( '' === $declared ) {
+            $declared = trim($this->attr($element, 'data-wp-layout'));
+        }
+
+        if ( '' !== $declared ) {
+            $decoded = json_decode($declared, true);
+            $type = is_array($decoded) ? (string) ($decoded['type'] ?? '') : $declared;
+            if ( in_array($type, array( 'constrained', 'flex', 'flow', 'grid' ), true) ) {
+                return array( 'type' => $type );
+            }
+        }
+
+        $style = strtolower('' !== trim($mergedStyle) ? $mergedStyle : $this->attr($element, 'style'));
+        if ( preg_match('/(?:^|;)\s*display\s*:\s*(inline-)?flex\b/', $style) ) {
+            return array( 'type' => 'flex' );
+        }
+        if ( preg_match('/(?:^|;)\s*display\s*:\s*(inline-)?grid\b/', $style) ) {
+            return array( 'type' => 'grid' );
+        }
+
+        if ( $this->hasGridLikeClass($element) && 1 < $this->cardLikeChildCount($element) ) {
+            return array( 'type' => 'grid' );
+        }
+
+        return array();
+    }
+
+    private function hasGridLikeClass(DOMElement $element): bool
+    {
+        $className = strtolower($this->attr($element, 'class'));
+        return (bool) preg_match('/(?:^|[\s_-])(?:cards|features|services|providers|testimonials|resources|posts|projects|stats|badges|grid|grid-[0-9]+|tiles|collection|gallery)(?:$|[\s_-])/', $className);
+    }
+}
